requirements.txt
discord.py==1.7.3
youtube_dl==2021.6.6
pycoingecko==2.2.0
PyNaCl==1.5.0
python-dotenv==0.21.0